Move file_downloaders, update references
