+++ utils/setup_gui/Base_Scripts/Scrapers/list_pdf_scrapers/single_pdf_scraper.py
@@ -7,7 +7,7 @@ import mimetypes
 p = Path(__file__).resolve().parents[3]
 sys.path.insert(1, str(p))
 
-from common.utils import get_pdf
+from utils.file_downloaders.downloaders import get_pdf
 
 
 def single_pdf_scraper(
